requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
streamlit
requests
